refactor(sdk): extract discovery + pubkey fetch to _client/_discovery.py
PR 3/N of the SDK refactor split (see #702 WebSocket and #706 Guardian
for the mixin template). Moves agent discovery and pubkey lookup
helpers out of cullis_sdk/client.py into a dedicated mixin so the
client.py shrink continues.

What moves:

- cullis_sdk/_client/_discovery.py — new module exposing
    * PubkeyFetchError (public, re-exported from cullis_sdk and
      cullis_sdk.client for back-compat)
    * _PUBKEY_CACHE_TTL = 300 (now scoped here, only used by the
      pubkey fetchers)
    * _DiscoveryMixin with:
        discover
        list_peers
        get_agent_public_key
        get_agent_public_key_via_egress
        _fetch_pubkey_proxy_then_broker

- cullis_sdk/client.py — CullisClient now inherits
  (_DiscoveryMixin, _GuardianMixin, _WebSocketMixin) in alphabetical
  order. The five methods, their section comment block, the module
  constant and the PubkeyFetchError class definition are removed.
  PubkeyFetchError is re-imported from the new canonical location so
  ``from cullis_sdk.client import PubkeyFetchError`` keeps working
  (existing tests + connector inbox poller rely on it).

- cullis_sdk/__init__.py — PubkeyFetchError import path updated to the
  new canonical location. Public ``cullis_sdk.PubkeyFetchError`` and
  ``cullis_sdk.WebSocketConnection`` remain reachable unchanged.

The host-class ``_pubkey_cache`` attribute (initialised in
CullisClient.__init__ and in every factory classmethod that bypasses
__init__) stays in client.py — it is not a Discovery concern, only
state that Discovery happens to consume.

Public surface unchanged. No logic change — pure code movement.
