+++ cullis_sdk/__init__.py
@@ -15,7 +15,8 @@ Usage::
         client.send(session_id, "myorg::agent", {"text": "Hello"}, agents[0].agent_id)
 """
 
-from cullis_sdk.client import CullisClient, PubkeyFetchError
+from cullis_sdk.client import CullisClient
+from cullis_sdk._client._discovery import PubkeyFetchError
 from cullis_sdk._client._websocket import WebSocketConnection
 from cullis_sdk.dpop import DpopKey
 from cullis_sdk.types import AgentInfo, SessionInfo, InboxMessage, RfqResult, RfqQuote
